Move bletia instance to KonashiManager
